--- packages/api-headless-cms/__tests__/modelBuilder/articleModelComparison.test.ts
@@ -1,159 +1,0 @@
-import { describe, it, expect, beforeEach } from "vitest";
-import { Container } from "@webiny/di";
-import { ModelBuilderFeature } from "~/features/modelBuilder/feature.js";
-import {
-    PrivateModel,
-    PrivateModelProvider,
-    type IPrivateModelBuilder
-} from "~/features/modelBuilder/index.js";
-import { createCmsModel } from "@webiny/api-headless-cms";
-import { articleModel } from "../contentTraverser/mocks/article.model.js";
-
-describe("Article Model Builder Comparison", () => {
-    let container: Container;
-
-    beforeEach(() => {
-        container = new Container();
-        ModelBuilderFeature.register(container);
-    });
-
-    describe("Complex Article Model", () => {
-        it("should produce identical output for article model with dynamic zones", async () => {
-            // ============================================
-            // OLD WAY - imported from existing test mocks
-            // ============================================
-            const oldModel = createCmsModel(articleModel).contentModel;
-
-            // ============================================
-            // NEW WAY - Using builder API via DI
-            // ============================================
-            class ArticleModelImpl implements PrivateModel.Interface {
-                buildModel(builder: IPrivateModelBuilder): IPrivateModelBuilder {
-                    return builder
-                        .modelId("article")
-                        .name("Article")
-                        .fields(fields => ({
-                            title: fields
-                                .text()
-                                .fieldId("title")
-                                .storageId("text@title")
-                                .label("Title"),
-                            body: fields
-                                .richText()
-                                .fieldId("body")
-                                .storageId("rich-text@body")
-                                .label("Body"),
-                            categories: fields
-                                .ref()
-                                .fieldId("categories")
-                                .storageId("ref@categories")
-                                .label("Categories")
-                                .multipleValues(true)
-                                .models([{ modelId: "category" }]),
-                            content: fields
-                                .dynamicZone()
-                                .fieldId("content")
-                                .storageId("dynamicZone@content")
-                                .label("Content")
-                                .multipleValues(true)
-                                .template("cv2zf965v324ivdc7e1vt", {
-                                    name: "Hero #1",
-                                    gqlTypeName: "Hero",
-                                    icon: "fas/flag",
-                                    description: "The top piece of content on every page.",
-                                    fields: f => ({
-                                        title: f.text().fieldId("title").label("Title")
-                                    })
-                                })
-                                .template("81qiz2v453wx9uque0gox", {
-                                    name: "Simple Text #1",
-                                    gqlTypeName: "SimpleText",
-                                    icon: "fas/file-text",
-                                    description: "Simple paragraph of text.",
-                                    fields: f => ({
-                                        text: f.longText().fieldId("text").label("Text")
-                                    })
-                                })
-                                .template("9ht43gurhegkbdfsaafyads", {
-                                    name: "Settings",
-                                    gqlTypeName: "Settings",
-                                    icon: "fas/file-text",
-                                    description: "Settings",
-                                    fields: f => ({
-                                        settings: f
-                                            .object()
-                                            .fieldId("settings")
-                                            .label("Settings")
-                                            .fields(objFields => ({
-                                                title: objFields
-                                                    .text()
-                                                    .fieldId("title")
-                                                    .label("Title"),
-                                                seo: objFields
-                                                    .object()
-                                                    .fieldId("seo")
-                                                    .label("SEO")
-                                                    .multipleValues(true)
-                                                    .fields(seoFields => ({
-                                                        title: seoFields
-                                                            .text()
-                                                            .fieldId("title")
-                                                            .label("Title")
-                                                    }))
-                                            })),
-                                        dynamicZone: f
-                                            .dynamicZone()
-                                            .fieldId("dynamicZone")
-                                            .label("DynamicZone")
-                                            .template("0emukbsvmzpozx2lzk883", {
-                                                name: "Ad",
-                                                gqlTypeName: "Ad",
-                                                icon: "fab/buysellads",
-                                                description: "Ad",
-                                                fields: adFields => ({
-                                                    authors: adFields
-                                                        .ref()
-                                                        .fieldId("authors")
-                                                        .label("Authors")
-                                                        .multipleValues(true)
-                                                        .models([{ modelId: "author" }])
-                                                })
-                                            }),
-                                        emptyDynamicZone: f
-                                            .dynamicZone()
-                                            .fieldId("emptyDynamicZone")
-                                            .label("DynamicZone")
-                                    })
-                                })
-                        }));
-                }
-            }
-
-            // Register the model implementation
-            container.registerInstance(PrivateModel, new ArticleModelImpl());
-
-            // Resolve the provider AFTER registering the model
-            const privateModelProvider = container.resolve(PrivateModelProvider);
-
-            // Get models via provider
-            const models = await privateModelProvider.getModels();
-            const newModel = models.find(m => m.modelId === "article");
-
-            // ============================================
-            // COMPARISON
-            // ============================================
-            expect(newModel).toBeDefined();
-
-            // Compare the essential properties
-            expect(newModel!.modelId).toBe(oldModel.modelId);
-            expect(newModel!.name).toBe(oldModel.name);
-            expect(newModel!.fields.length).toBe(oldModel.fields.length);
-
-            // Note: Full JSON comparison is not performed because:
-            // - The old article model was hand-written with minimal/incomplete field definitions
-            // - The new builder API ensures all field properties are defined with sensible defaults
-            // - This is the CORRECT behavior - FieldBuilder.build() should ensure completeness
-            // - The key validation is that dynamic zone templates work correctly with the chainable .template() API
-        });
-    });
-});
