+++ packages/api-headless-cms/src/features/modelBuilder/fields/DynamicZoneFieldType.ts
@@ -56,7 +56,10 @@ class DynamicZoneFieldBuilder
         const fieldBuilders = config.fields(this.registry);
         const fields: any[] = [];
 
-        for (const [, fieldBuilder] of Object.entries(fieldBuilders)) {
+        for (const [key, fieldBuilder] of Object.entries(fieldBuilders)) {
+            // Automatically set the fieldId from the object key
+            // This ensures the key and fieldId are always in sync
+            fieldBuilder.fieldId(key);
             fields.push((fieldBuilder as any).build());
         }
